보드쪽 손대는중.

+++ youthmap/src/main/java/com/example/demo/model/Comment.java
@@ -15,5 +15,6 @@ public class Comment {
     private String commContent;  // 댓글 내용
     private Date commDate;       // 작성 날짜
     private Date commUpdate;     // 수정 날짜 (수정 시 사용)
-    private int memNo;
+    private Long memNo;          // 회원 번호 (MemberModel과 일관성)
+
 }
